refactor: split floating back button behavior
