+++ src/components/FloatingBackButton.tsx
@@ -1,224 +1,19 @@
-﻿import { useEffect, useMemo, useRef, useState, type MouseEvent, type PointerEvent as ReactPointerEvent } from 'react';
-
-const BUTTON_SIZE = 46;
-const EDGE_PADDING = 12;
-const DESKTOP_BOTTOM_PADDING = 120;
-const MOBILE_SHEET_OFFSET = 180;
-const TOUCH_DRAG_DELAY_MS = 260;
-
-interface Position {
-  x: number;
-  y: number;
-}
-
-interface Bounds {
-  minX: number;
-  maxX: number;
-  minY: number;
-  maxY: number;
-}
-
-function getPlacementBounds(): Bounds | null {
-  if (typeof window === 'undefined') {
-    return null;
-  }
-
-  const shell = document.querySelector<HTMLElement>('.phone-shell');
-  if (shell) {
-    const rect = shell.getBoundingClientRect();
-    return {
-      minX: rect.left + EDGE_PADDING,
-      maxX: rect.right - BUTTON_SIZE - EDGE_PADDING,
-      minY: rect.top + EDGE_PADDING,
-      maxY: rect.bottom - BUTTON_SIZE - EDGE_PADDING,
-    };
-  }
-
-  return {
-    minX: EDGE_PADDING,
-    maxX: Math.max(EDGE_PADDING, window.innerWidth - BUTTON_SIZE - EDGE_PADDING),
-    minY: EDGE_PADDING,
-    maxY: Math.max(EDGE_PADDING, window.innerHeight - BUTTON_SIZE - DESKTOP_BOTTOM_PADDING),
-  };
-}
-
-function getDefaultPosition(): Position {
-  if (typeof window === 'undefined') {
-    return { x: EDGE_PADDING, y: EDGE_PADDING };
-  }
-
-  const bounds = getPlacementBounds();
-  if (!bounds) {
-    return { x: EDGE_PADDING, y: EDGE_PADDING };
-  }
-
-  const isMobileViewport = window.innerWidth <= 640;
-  return clampPosition({
-    x: bounds.maxX,
-    y: isMobileViewport ? bounds.maxY - MOBILE_SHEET_OFFSET : bounds.maxY - DESKTOP_BOTTOM_PADDING,
-  });
-}
-
-function clampPosition(position: Position) {
-  const bounds = getPlacementBounds();
-  if (!bounds) {
-    return position;
-  }
-
-  return {
-    x: Math.min(Math.max(position.x, bounds.minX), bounds.maxX),
-    y: Math.min(Math.max(position.y, bounds.minY), bounds.maxY),
-  };
-}
+﻿import { useFloatingBackButton } from './floating-back-button/useFloatingBackButton';
 
 interface FloatingBackButtonProps {
   onNavigateBack: () => void;
 }
 
 export function FloatingBackButton({ onNavigateBack }: FloatingBackButtonProps) {
-  const [position, setPosition] = useState<Position>(getDefaultPosition);
-  const [isDragging, setIsDragging] = useState(false);
-  const handledTapRef = useRef(false);
-  const touchTimerRef = useRef<number | null>(null);
-  const dragStateRef = useRef({
-    pointerId: -1,
-    startX: 0,
-    startY: 0,
-    originX: 0,
-    originY: 0,
-    hasMoved: false,
-    dragEnabled: false,
-  });
-
-  useEffect(() => {
-    if (typeof window === 'undefined') {
-      return undefined;
-    }
-
-    function handleResize() {
-      setPosition((current) => clampPosition(current));
-    }
-
-    window.addEventListener('resize', handleResize);
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
-
-  useEffect(() => () => {
-    if (touchTimerRef.current !== null) {
-      window.clearTimeout(touchTimerRef.current);
-      touchTimerRef.current = null;
-    }
-  }, []);
-
-  const style = useMemo(
-    () => ({ left: position.x, top: position.y, right: 'auto', bottom: 'auto' }),
-    [position],
-  );
-
-  function clearTouchTimer() {
-    if (touchTimerRef.current !== null) {
-      window.clearTimeout(touchTimerRef.current);
-      touchTimerRef.current = null;
-    }
-  }
-
-  function finishDrag(pointerId: number) {
-    if (dragStateRef.current.pointerId !== pointerId) {
-      return;
-    }
-
-    clearTouchTimer();
-    dragStateRef.current.pointerId = -1;
-    dragStateRef.current.dragEnabled = false;
-    setIsDragging(false);
-  }
-
-  function handlePointerDown(event: ReactPointerEvent<HTMLButtonElement>) {
-    if (event.button !== 0) {
-      return;
-    }
-
-    handledTapRef.current = false;
-    dragStateRef.current = {
-      pointerId: event.pointerId,
-      startX: event.clientX,
-      startY: event.clientY,
-      originX: position.x,
-      originY: position.y,
-      hasMoved: false,
-      dragEnabled: event.pointerType === 'mouse',
-    };
-
-    if (event.pointerType !== 'mouse') {
-      clearTouchTimer();
-      touchTimerRef.current = window.setTimeout(() => {
-        if (dragStateRef.current.pointerId === event.pointerId) {
-          dragStateRef.current.dragEnabled = true;
-          setIsDragging(true);
-        }
-      }, TOUCH_DRAG_DELAY_MS);
-    } else {
-      setIsDragging(true);
-    }
-
-    event.currentTarget.setPointerCapture(event.pointerId);
-  }
-
-  function handlePointerMove(event: ReactPointerEvent<HTMLButtonElement>) {
-    if (dragStateRef.current.pointerId !== event.pointerId) {
-      return;
-    }
-
-    const deltaX = event.clientX - dragStateRef.current.startX;
-    const deltaY = event.clientY - dragStateRef.current.startY;
-    if (Math.abs(deltaX) > 4 || Math.abs(deltaY) > 4) {
-      dragStateRef.current.hasMoved = true;
-    }
-
-    if (!dragStateRef.current.dragEnabled) {
-      return;
-    }
-
-    event.preventDefault();
-    setIsDragging(true);
-    setPosition(
-      clampPosition({
-        x: dragStateRef.current.originX + deltaX,
-        y: dragStateRef.current.originY + deltaY,
-      }),
-    );
-  }
-
-  function handlePointerUp(event: ReactPointerEvent<HTMLButtonElement>) {
-    const shouldNavigate =
-      dragStateRef.current.pointerId === event.pointerId &&
-      !dragStateRef.current.hasMoved &&
-      !dragStateRef.current.dragEnabled;
-    finishDrag(event.pointerId);
-    if (shouldNavigate) {
-      handledTapRef.current = true;
-      onNavigateBack();
-    }
-  }
-
-  function handlePointerCancel(event: ReactPointerEvent<HTMLButtonElement>) {
-    finishDrag(event.pointerId);
-  }
-
-  function handleClick(event: MouseEvent<HTMLButtonElement>) {
-    if (handledTapRef.current) {
-      handledTapRef.current = false;
-      event.preventDefault();
-      return;
-    }
-
-    if (dragStateRef.current.hasMoved || isDragging) {
-      event.preventDefault();
-      return;
-    }
-
-    onNavigateBack();
-  }
+  const {
+    isDragging,
+    style,
+    handlePointerDown,
+    handlePointerMove,
+    handlePointerUp,
+    handlePointerCancel,
+    handleClick,
+  } = useFloatingBackButton({ onNavigateBack });
 
   return (
     <button
