+++ src/components/floating-back-button/constants.ts
@@ -1,0 +1,5 @@
+export const BUTTON_SIZE = 46;
+export const EDGE_PADDING = 12;
+export const DESKTOP_BOTTOM_PADDING = 120;
+export const MOBILE_SHEET_OFFSET = 180;
+export const TOUCH_DRAG_DELAY_MS = 260;
